Add files via upload
The research on the gold trade exports has been conducted with the various techniques and 
methods such as below:
3.1) Data Preprocessing:
The source data has been processed based on the data wrangling techniques such as below:
3.1.1) Data Availability:
The source data for this research has been taken from the UN ComTrade database. The data has 
been obtained from source through the API by passing the below parameters:
Freqcode = M #Monthly
ClsCode = HS #HS Classification
Period = 2017 to 2023
ReporterCode = 404, 710, 818 #Kenya, South Africa, Egypt
CmdCode = 7108, 710811, 710812, 710813, 710820 #Gold
FlowCode = X #Export
3.1.2) Data Analysis:
Based on the “Mapping Document.xlsx” workbook [mentioned in Appendix 1], data analysis has 
been conducted to the access data via API calls and the data has been filtered based on given criteria 
mentioned under “Assumptions” column in “Mapping” Sheet. The raw source data extracted via API is 
archived in the data lake table named as “GOLD_TRADE_DL”
3.1.3) Data Transformation:
For certain attributes with similar values, data transformation has been performed so that the final 
records contain only non-zero/non-null values for the mandatory columns and sent for further 
processing. Below are the conditions listed for the transformation:
i. Primary Value: This attribute defines the trade value in dollars. The records with the trade 
value such as blank, or null values are considered as invalid. In such conditions, the 
UE, Potsdam Group 3 PS24 Data Science and Business Pg 6
corresponding value is copied directly from the FOB_VALUE, as FOB_VALUE contains exact 
export values.
ii. Net Weight: This attribute defines the trade volume in kilograms. The records with the trade 
volume such as blank, or null values are treated as invalid. In such cases, data has been
compared with Qty and AltQty attributes. If the value is in grams, data has been converted 
and then copied to NetWeight in Kilograms as we are dealing with only grams and kilograms 
data from the source.
3.1.4) Data Mean Substitution:
For certain records, we have identified that the data has been missing or as 0 for the required 
attributes. In such cases, to avoid loss of data, data has been filled with mean values from the other 
months for the reporters which makes the data consistent while plotting for visualization.
3.1.5) Data Outliers Handling:
As we are dealing with the real-time data with combination of reported and aggregated values, 
there is a high possibility that data may fall on invalid range which makes the visualization inconsistent. 
To avoid such issues, z-score outliers has been performed by calculating the relation between the 
important attributes and then those records are removed which makes the data ready for visualization.
3.1.6) Data Validation:
After analyzing the data, there are lot of duplicates found on basis of aggregated values. To 
clean those duplicate sum values, data validation has been performed using “Data Validation.xlsx” 
[mentioned in Appendix 2]. Records that fail to satisfy the conditions are rejected and stored in the 
reject table named as “GOLD_TRADE_DATA_REJECT “. Records that met the conditions are stored in the 
data warehouse table named as “GOLD_TRADE”, with the final columns listed in the requirement 
document.
3.1.7) Data Readiness:
To visualize the data, not all the attributes mentioned in the requirement specifications are 
required. Hence, unwanted attributes are removed and are stored in the data mart table named as 
“GOLD_TRADE_DATA_MART”.
UE, Potsdam Group 3 PS24 Data Science and Business Pg 7
3.2) Other Data Sources:
Historical standard gold prices for comparison are collected from world bank [7], then converted 
from oz to the kilograms and stored in the table named “GOLD_PRICE”.
Information that are related to the BRICS countries are sourced from Wikipedia [8] and then 
updated with Partner data taken from ComTrade, stored in the table named “BRICS_CONT”